+++ public/app/store/configureStore.ts
@@ -10,7 +10,7 @@ import dashboardReducers from 'app/features/dashboard/state/reducers';
 import pluginReducers from 'app/features/plugins/state/reducers';
 import dataSourcesReducers from 'app/features/datasources/state/reducers';
 import usersReducers from 'app/features/users/state/reducers';
-import appNotificationReducers from 'app/core/components/AppNotifications/state/reducers';
+import appNotificationReducers from 'app/core/reducers/appNotification';
 
 const rootReducers = {
   ...sharedReducers,
