+++ public/views/index.template.html
@@ -200,7 +200,7 @@
 
   <grafana-app class="grafana-app" ng-cloak>
     <sidemenu class="sidemenu"></sidemenu>
-    <page-alert-list class="page-alert-list"></page-alert-list>
+    <app-notifications-list class="page-alert-list"></app-notifications-list>
 
 
     <div class="main-view">
